Implement facebook integration
Functionality included in the SocialMediaHandler class. Also move files from core to android so the handlers can have access to the AndroidManifest. The AndroidManifest defines the FacebookActivity, which is necessary to open up the fb login/share video screens.
