+++ code/android/build.gradle
@@ -18,7 +18,7 @@ android {
     }
     defaultConfig {
         applicationId "com.ece454.gotl"
-        minSdkVersion 9
+        minSdkVersion 15
         targetSdkVersion 28
         versionCode 1
         versionName "1.0"
