+++ code/android/res/values/strings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
     <string name="app_name">goose-on-the-loose</string>
-
+    <string name="facebook_app_id">2402208109839122</string>
+    <string name="fb_login_protocol_scheme">fb2402208109839122</string>
 </resources>
