+++ code/android/src/com/ece454/gotl/AndroidLauncher.java
@@ -4,15 +4,19 @@ import android.os.Bundle;
 
 import com.badlogic.gdx.backends.android.AndroidApplication;
 import com.badlogic.gdx.backends.android.AndroidApplicationConfiguration;
-import com.ece454.gotl.GotlGame;
+
+import handlers.SocialMediaHandler;
 
 public class AndroidLauncher extends AndroidApplication {
+    private GotlGame game;
 	@Override
 	protected void onCreate (Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		AndroidApplicationConfiguration config = new AndroidApplicationConfiguration();
 		config.useAccelerometer = false;
 		config.useCompass = false;
-		initialize(new GotlGame(), config);
+		game = new GotlGame();
+		game.setSocialMediaHandler(new SocialMediaHandler(this));
+		initialize(game, config);
 	}
 }
