+++ code/android/src/com/ece454/gotl/Bounds.java
@@ -13,6 +13,5 @@ public class Bounds {
         bodyDef.type = BodyDef.BodyType.StaticBody;
         body = world.createBody(bodyDef);
         body.createFixture(shape, DENSITY).setUserData(this);
-//        shape.dispose();
     }
 }
