+++ code/android/src/com/ece454/gotl/Ground.java
@@ -13,6 +13,5 @@ public class Ground {
         bodyDef.type = BodyDef.BodyType.StaticBody;
         body = world.createBody(bodyDef);
         body.createFixture(shape, DENSITY).setUserData(this);
-//        shape.dispose();
     }
 }
