+++ code/android/src/com/ece454/gotl/MapParser.java
