+++ code/android/src/handlers/WorldManager.java
@@ -2,7 +2,6 @@ package handlers;
 
 import com.badlogic.gdx.maps.tiled.TiledMap;
 import com.badlogic.gdx.math.Vector2;
-import com.badlogic.gdx.physics.box2d.BodyDef;
 import com.badlogic.gdx.physics.box2d.World;
 import com.ece454.gotl.MapParser;
 import com.ece454.gotl.WorldContactListener;
